merged: InceptionVault_S, InceptionVault_E
