+++ projects/vaults/contracts/interfaces/symbiotic-vault/symbiotic-core/ISymbioticVault.sol
