+++ projects/vaults/contracts/vaults/EigenLayer/facets/ERC4626Facet/ERC4626Facet_EL_E1.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.27;
+pragma solidity 0.8.24;
 
 import {ERC4626Facet_EL} from "./ERC4626Facet_EL.sol";
