+++ projects/vaults/contracts/vaults/EigenLayer/facets/EigenSetterFacet.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.27;
+pragma solidity 0.8.24;
 
 import {Address} from "@openzeppelin/contracts/proxy/beacon/BeaconProxy.sol";
