+++ projects/vaults/test/helpers/utils.js
@@ -22,9 +22,8 @@ const calculateRatio = async (vault, token) => {
   }
 
   if (denominator === 0n || totalSupply === 0n) {
-    const ratio = e18;
-    // console.log(`Current ratio is:\t\t\t\t${ratio.format()}`);
-    return ratio;
+    console.log("iToken supply is 0, so the ration is going to be 1e18");
+    return e18;
   }
 
   const ratio = (totalSupply * e18) / denominator;
@@ -102,7 +101,7 @@ const randomBI = (length) => {
 const randomBIMax = (max) => {
   let random = 0n;
   if (max > 0n) {
-    random += BigInt(Math.random() * Number(max));
+    random += BigInt(Math.floor(Math.random() * Number(max)));
   }
   return random;
 };
@@ -113,6 +112,8 @@ const randomAddress = () => ethers.Wallet.createRandom().address;
 const format = (bi) => bi.toLocaleString("de-DE");
 
 const e18 = 1000_000_000_000_000_000n;
+const e9 = 1000_000_000n;
+const zeroWithdrawalData = [ethers.ZeroAddress, ethers.ZeroAddress, ethers.ZeroAddress, 0, 1, [ethers.ZeroAddress], [0]];
 
 const day = 86400n;
